+++ Invoices-App-Client/src/components/Layout/index.jsx
@@ -1,10 +1,12 @@
 import Navbar from '../Navbar'
 
-export default function Layout({ children }) {
+export default function AuthLayout({ children }) {
   return (
-    <>
+    <section className="w-full flex flex-col justify-center items-center">
       <Navbar />
-      {children}
-    </>
+      <article className="w-full flex flex-col items-center justify-center gap-10 lg:flex-row lg:max-w-7xl pt-8">
+        {children}
+      </article>
+    </section>
   )
 }
