+++ Invoices-App-Client/src/components/Navbar/index.jsx
@@ -1,34 +1,19 @@
-import { Link } from 'react-router-dom'
-import logo from './aimedge_logo.jpg'
+import { useNavigate } from 'react-router-dom'
+import { useCallback } from 'react'
 
 export default function Navbar() {
-  const navLinks = [
-    { path: '/', name: 'Home' },
-    { path: '/login', name: 'Login' },
-    { path: '/register', name: 'Register' },
-  ]
+  const goTo = useNavigate()
+  const handleClick = useCallback(() => goTo('/'), [goTo])
+
   return (
-    <nav className=" bg-white w-[100%] justify-center absolute items-center max-w-[100%] flex flex-col h-20 shadow-md">
-      <div className="flex flex-row w-[100%] max-w-[800px] justify-between items-center">
-        <div className="w-11 h-11">
-          <img
-            alt="AIMED EDGE LOGO"
-            src={logo}
-            className="w-[100%] h-[100%] object-contain"
-          />
-        </div>
-        <div className="flex flex-col">
-          <ul className="flex flex-row justify-between gap-2">
-            {navLinks.map((link) => (
-              <li
-                className="list-none text-blue-gray-300 font-semibold text-[16px] cursor-pointer hover:text-blue-gray-100"
-                key={link.name}
-              >
-                <Link to={link.path}>{link.name}</Link>
-              </li>
-            ))}
-          </ul>
-        </div>
+    <nav className="w-full bg-transparent flex flex-col py-4">
+      <div className="w-full max-w-[1366px] flex flex-row justify-between items-center self-center">
+        <img
+          src={'/app-logo.svg'}
+          alt="invoices app logo"
+          className="w-16 cursor-pointer"
+          onClick={handleClick}
+        />
       </div>
     </nav>
   )
