+++ Invoices-App-Client/src/pages/Register/index.jsx
@@ -1,12 +1,5 @@
-import RegisterForm from '../../components/Forms/RegisterForm'
-import Layout from '../../components/Layout'
+import RegisterScreen from '../../screens/Register'
 
 export default function Register() {
-  return (
-    <Layout>
-      <section className="w-full h-full flex flex-col items-center justify-center">
-        <RegisterForm />
-      </section>
-    </Layout>
-  )
+  return <RegisterScreen />
 }
