+++ Invoices-App-Client/src/screens/Login/index.jsx
@@ -1,0 +1,10 @@
+import LoginForm from '../../components/Forms/LoginForm'
+import AuthLayout from '../../components/Layout'
+
+export default function LoginScreen() {
+  return (
+    <AuthLayout>
+      <LoginForm />
+    </AuthLayout>
+  )
+}
