+++ Invoices-App-Client/src/screens/Register/index.jsx
@@ -1,0 +1,10 @@
+import AuthLayout from '../../components/Layout'
+import RegisterForm from '../../components/Forms/RegisterForm'
+
+export default function RegisterScreen() {
+  return (
+    <AuthLayout>
+      <RegisterForm />
+    </AuthLayout>
+  )
+}
